Move tile worker code to separate source files
