+++ include/tile_worker.h
@@ -1,0 +1,8 @@
+#ifndef _TILE_WORKER_H
+#define _TILE_WORKER_H
+
+#include "shared_data.h"
+
+int outputProc(uint threadId, class SharedData *sharedData);
+
+#endif //_TILE_WORKER_H
